py/app/requirements.txt
Flask-SQLAlchemy==2.2
flask-restful
